Add measurementsSavingService
