--- AirCasting/BluetoothSession/AirBeamMeasurement.swift
@@ -1,9 +1,0 @@
-// Created by Lunar on 14/11/2022.
-//
-
-import Foundation
-
-struct AirBeamMeasurement {
-    let device: NewBluetoothManager.BluetoothDevice
-    let measurementStream: ABMeasurementStream
-}
