+++ AirCasting/BluetoothSession/MobilePeripheralSessionManager.swift
@@ -5,24 +5,23 @@ import Foundation
 import CoreLocation
 import Resolver
 
-
 //Change name to saver
 class MobilePeripheralSessionManager {
-    class PeripheralMeasurementTimeLocationManager {
-        @Injected private var locationTracker: LocationTracker
-        
-        private(set) var collectedValuesCount: Int = 5
-        private(set) var currentTime: Date = DateBuilder.getFakeUTCDate()
-        private(set) var currentLocation: CLLocationCoordinate2D? = .undefined
-        
-        func startNewValuesRound(locationless: Bool) {
-            currentLocation = !locationless ? locationTracker.location.value?.coordinate : .undefined
-            currentTime = DateBuilder.getFakeUTCDate()
-            collectedValuesCount = 0
-        }
-        
-        func incrementCounter() { collectedValuesCount += 1 }
-    }
+//    class PeripheralMeasurementTimeLocationManager {
+//        @Injected private var locationTracker: LocationTracker
+//
+//        private(set) var collectedValuesCount: Int = 5
+//        private(set) var currentTime: Date = DateBuilder.getFakeUTCDate()
+//        private(set) var currentLocation: CLLocationCoordinate2D? = .undefined
+//
+//        func startNewValuesRound(locationless: Bool) {
+//            currentLocation = !locationless ? locationTracker.location.value?.coordinate : .undefined
+//            currentTime = DateBuilder.getFakeUTCDate()
+//            collectedValuesCount = 0
+//        }
+//
+//        func incrementCounter() { collectedValuesCount += 1 }
+//    }
     
     
     // Used to protect screen when bt session is recording
@@ -34,7 +33,7 @@ class MobilePeripheralSessionManager {
     @Injected private var btManager: BluetoothConnectionHandler
     
     // Used for adding right time and location to all 5 streams
-    private var peripheralMeasurementManager = PeripheralMeasurementTimeLocationManager()
+//    private var peripheralMeasurementManager = PeripheralMeasurementTimeLocationManager()
     var activeMobileSession: MobileSession?
     
     func startRecording(session: Session, device: NewBluetoothManager.BluetoothDevice) {
@@ -65,14 +64,14 @@ class MobilePeripheralSessionManager {
         }
     }
     
-    func handlePeripheralMeasurement(_ measurement: AirBeamMeasurement) {
-        guard activeMobileSession?.device == measurement.device else { return }
-        
-        if peripheralMeasurementManager.collectedValuesCount == 5 { peripheralMeasurementManager.startNewValuesRound(locationless: activeMobileSession!.session.locationless) }
-        
-        updateStreams(stream: measurement.measurementStream, sessionUUID: activeMobileSession!.session.uuid, location: peripheralMeasurementManager.currentLocation, time: peripheralMeasurementManager.currentTime)
-        peripheralMeasurementManager.incrementCounter()
-    }
+//    func handlePeripheralMeasurement(_ measurement: AirBeamMeasurement) {
+//        guard activeMobileSession?.device == measurement.device else { return }
+//
+//        if peripheralMeasurementManager.collectedValuesCount == 5 { peripheralMeasurementManager.startNewValuesRound(locationless: activeMobileSession!.session.locationless) }
+//
+//        updateStreams(stream: measurement.measurementStream, sessionUUID: activeMobileSession!.session.uuid, location: peripheralMeasurementManager.currentLocation, time: peripheralMeasurementManager.currentTime)
+//        peripheralMeasurementManager.incrementCounter()
+//    }
     
     func activeSessionInProgressWith(_ device: NewBluetoothManager.BluetoothDevice) -> Bool {
         activeMobileSession?.device == device
@@ -151,36 +150,37 @@ class MobilePeripheralSessionManager {
         }
     }
 
-    private func updateStreams(stream: ABMeasurementStream, sessionUUID: SessionUUID, location: CLLocationCoordinate2D?, time: Date) {
-        measurementStreamStorage.accessStorage { storage in
-            do {
-                let existingStreamID = try storage.existingMeasurementStream(sessionUUID, name: stream.sensorName)
-                guard let id = existingStreamID else {
-                    let streamId = try self.createSessionStream(stream, sessionUUID, storage: storage)
-                    try storage.addMeasurementValue(stream.measuredValue, at: location, toStreamWithID: streamId, on: time)
-                    return
-                }
-                try storage.addMeasurementValue(stream.measuredValue, at: location, toStreamWithID: id, on: time)
-            } catch {
-                Log.error("Error saving value from peripheral: \(error)")
-            }
-        }
-    }
+//    private func updateStreams(stream: ABMeasurementStream, sessionUUID: SessionUUID, location: CLLocationCoordinate2D?, time: Date) {
+//        measurementStreamStorage.accessStorage { storage in
+//            do {
+//                let existingStreamID = try storage.existingMeasurementStream(sessionUUID, name: stream.sensorName)
+//                guard let id = existingStreamID else {
+//                    let streamId = try self.createSessionStream(stream, sessionUUID, storage: storage)
+//                    try storage.addMeasurementValue(stream.measuredValue, at: location, toStreamWithID: streamId, on: time)
+//                    return
+//                }
+//                try storage.addMeasurementValue(stream.measuredValue, at: location, toStreamWithID: id, on: time)
+//            } catch {
+//                Log.error("Error saving value from peripheral: \(error)")
+//            }
+//        }
+//    }
 
-    private func createSessionStream(_ stream: ABMeasurementStream, _ sessionUUID: SessionUUID, storage: HiddenCoreDataMeasurementStreamStorage) throws -> MeasurementStreamLocalID {
-        let sessionStream = MeasurementStream(id: nil,
-                                              sensorName: stream.sensorName,
-                                              sensorPackageName: stream.packageName,
-                                              measurementType: stream.measurementType,
-                                              measurementShortType: stream.measurementShortType,
-                                              unitName: stream.unitName,
-                                              unitSymbol: stream.unitSymbol,
-                                              thresholdVeryHigh: Int32(stream.thresholdVeryHigh),
-                                              thresholdHigh: Int32(stream.thresholdHigh),
-                                              thresholdMedium: Int32(stream.thresholdMedium),
-                                              thresholdLow: Int32(stream.thresholdLow),
-                                              thresholdVeryLow: Int32(stream.thresholdVeryLow))
-
-        return try storage.saveMeasurementStream(sessionStream, for: sessionUUID)
-    }
+//    private func createSessionStream(_ stream: ABMeasurementStream, _ sessionUUID: SessionUUID, storage: HiddenCoreDataMeasurementStreamStorage) throws -> MeasurementStreamLocalID {
+//        let sessionStream = MeasurementStream(id: nil,
+//                                              sensorName: stream.sensorName,
+//                                              sensorPackageName: stream.packageName,
+//                                              measurementType: stream.measurementType,
+//                                              measurementShortType: stream.measurementShortType,
+//                                              unitName: stream.unitName,
+//                                              unitSymbol: stream.unitSymbol,
+//                                              thresholdVeryHigh: Int32(stream.thresholdVeryHigh),
+//                                              thresholdHigh: Int32(stream.thresholdHigh),
+//                                              thresholdMedium: Int32(stream.thresholdMedium),
+//                                              thresholdLow: Int32(stream.thresholdLow),
+//                                              thresholdVeryLow: Int32(stream.thresholdVeryLow))
+//
+//        return try storage.saveMeasurementStream(sessionStream, for: sessionUUID)
+//    }
 }
+
